修改：将GB2312编码的文件转为UTF-8-BOM编码，补齐注释

+++ RemoteCtrl/RemoteClient/RemoteClientDlg.cpp
@@ -581,7 +581,7 @@ void CRemoteClientDlg::threadWatchData()
 	{
 		if (GetTickCount64() - ulTick < 150)
 		{
-			Sleep(GetTickCount64() - ulTick);
+			Sleep(DWORD(GetTickCount64() - ulTick));
 		}
 
 		if (!m_bIsFull)//更新数据到缓存
